Added CORS, Added test index.html with fetch

+++ src/modules/todo/controllers/todo.controller.ts
@@ -13,7 +13,7 @@ export class TodoController {
 
   @Get(':id')
   async getOneAction(@Param('id') id: string): Promise<Todo> {
-    const todo = this.todoService.findOne(id);
+    const todo = await this.todoService.findOne(id);
     if (todo === undefined) {
       throw new HttpException(
         'Todo with id=' + id + 'not exists',
@@ -49,7 +49,14 @@ export class TodoController {
   }
 
   @Delete(':id')
-  deleteAction(@Param('id') id: string): Promise<void> {
+  async deleteAction(@Param('id') id: string): Promise<void> {
+    const todo = await this.todoService.findOne(id);
+    if (todo === undefined) {
+      throw new HttpException(
+        'Todo with id=' + id + 'not exists',
+        HttpStatus.NOT_FOUND
+        );
+    }
     return this.todoService.remove(id);
   }
 }
